Address review: convert EditTextInputAction to object, restructure injector (#6160)
