+++ testing/BUILD.bazel
@@ -104,6 +104,7 @@ TEST_DEPS = [
     "//testing/src/main/java/org/oppia/android/testing/espresso:text_input_action",
     "//testing/src/main/java/org/oppia/android/testing/platformparameter:test_module",
     "//testing/src/main/java/org/oppia/android/testing/robolectric:test_module",
+    "//testing/src/main/java/org/oppia/android/testing/threading:test_coroutine_dispatchers_injector",
     "//testing/src/main/java/org/oppia/android/testing/threading:test_module",
     "//testing/src/main/java/org/oppia/android/testing/time:test_module",
     "//third_party:androidx_lifecycle_lifecycle-livedata-ktx",
