+++ testing/src/main/java/org/oppia/android/testing/espresso/EditTextInputAction.kt
@@ -9,7 +9,6 @@ import androidx.test.espresso.action.ViewActions
 import androidx.test.espresso.action.ViewActions.typeText
 import org.hamcrest.Matcher
 import org.oppia.android.testing.threading.TestCoroutineDispatchersInjector
-import javax.inject.Inject
 
 /**
  * Action for inputting text into an EditText in a test infrastructure-specific way.
@@ -18,47 +17,45 @@ import javax.inject.Inject
  * 'android:digits' or other filters. See https://github.com/robolectric/robolectric/issues/5110
  * for specifics.
  */
-class EditTextInputAction @Inject constructor() {
-  companion object {
-    /**
-     * Returns a [ViewAction] that appends the specified string into the view targeted by the
-     * [ViewAction].
-     */
-    fun appendText(text: String): ViewAction = updateText(
-      text, baseAction = typeText(text), isAppend = true
-    )
+object EditTextInputAction {
+  /**
+   * Returns a [ViewAction] that appends the specified string into the view targeted by the
+   * [ViewAction].
+   */
+  fun appendText(text: String): ViewAction = updateText(
+    text, baseAction = typeText(text), isAppend = true
+  )
 
-    /**
-     * Returns a [ViewAction] that replaces the current text in the specified view with the specified
-     * string.
-     *
-     * Note that this should only be used over [appendText] in the following cases:
-     * 1. When there's existing text to first erase before adding new text
-     * 2. When Unicode text needs to be inputted (since otherwise Espresso will fail to type the text)
-     */
-    fun replaceText(text: String): ViewAction = updateText(
-      text, baseAction = ViewActions.replaceText(text), isAppend = false
-    )
+  /**
+   * Returns a [ViewAction] that replaces the current text in the specified view with the specified
+   * string.
+   *
+   * Note that this should only be used over [appendText] in the following cases:
+   * 1. When there's existing text to first erase before adding new text
+   * 2. When Unicode text needs to be inputted (since otherwise Espresso will fail to type the text)
+   */
+  fun replaceText(text: String): ViewAction = updateText(
+    text, baseAction = ViewActions.replaceText(text), isAppend = false
+  )
 
-    private fun updateText(text: String, baseAction: ViewAction, isAppend: Boolean): ViewAction {
-      return object : ViewAction {
-        override fun getDescription(): String = baseAction.description
+  private fun updateText(text: String, baseAction: ViewAction, isAppend: Boolean): ViewAction {
+    return object : ViewAction {
+      override fun getDescription(): String = baseAction.description
 
-        override fun getConstraints(): Matcher<View> = baseAction.constraints
+      override fun getConstraints(): Matcher<View> = baseAction.constraints
 
-        override fun perform(uiController: UiController?, view: View?) {
-          // Appending text only works on Robolectric, whereas Espresso needs to use typeText().
-          if (Build.FINGERPRINT.contains("robolectric", ignoreCase = true)) {
-            val editText = view as? EditText
-            if (isAppend) {
-              editText?.append(text)
-            } else {
-              editText?.setText(text)
-            }
-            TestCoroutineDispatchersInjector.getDispatcher().runCurrent()
+      override fun perform(uiController: UiController?, view: View?) {
+        // Appending text only works on Robolectric, whereas Espresso needs to use typeText().
+        if (Build.FINGERPRINT.contains("robolectric", ignoreCase = true)) {
+          val editText = view as? EditText
+          if (isAppend) {
+            editText?.append(text)
           } else {
-            baseAction.perform(uiController, view)
+            editText?.setText(text)
           }
+          TestCoroutineDispatchersInjector.getDispatcher().runCurrent()
+        } else {
+          baseAction.perform(uiController, view)
         }
       }
     }
