+++ testing/src/main/java/org/oppia/android/testing/threading/BUILD.bazel
@@ -40,9 +40,23 @@ kt_android_library(
     testonly = True,
     srcs = [
         "TestCoroutineDispatchers.kt",
-        "TestCoroutineDispatchersInjector.kt",
     ],
     visibility = ["//:oppia_testing_visibility"],
+)
+
+kt_android_library(
+    name = "test_coroutine_dispatchers_injector",
+    testonly = True,
+    srcs = [
+        "TestCoroutineDispatchersInjector.kt",
+    ],
+    visibility = [
+        "//testing:__pkg__",
+        "//testing/src/main/java/org/oppia/android/testing:__subpackages__",
+    ],
+    deps = [
+        ":test_coroutine_dispatchers",
+    ],
 )
 
 kt_android_library(
@@ -108,6 +122,7 @@ kt_android_library(
         ":espresso_impl",
         ":robolectric_impl",
         ":test_coroutine_dispatchers",
+        ":test_coroutine_dispatchers_injector",
         "//:dagger",
         "//utility/src/main/java/org/oppia/android/util/threading:annotations",
     ],
